Extract catalog version lifecycle
